Move types into matching files and adjust namespaces
