+++ src/Microsoft.TemplateEngine.Abstractions/IEngineEnvironmentSettings.cs
@@ -1,0 +1,13 @@
+namespace Microsoft.TemplateEngine.Abstractions
+{
+    public interface IEngineEnvironmentSettings
+    {
+        ISettingsLoader SettingsLoader { get; }
+
+        ITemplateEngineHost Host { get; }
+
+        IEnvironment Environment { get; }
+
+        IPathInfo Paths { get; }
+    }
+}
