+++ src/Microsoft.TemplateEngine.Abstractions/ISettingsLoader.cs
@@ -32,15 +32,4 @@ namespace Microsoft.TemplateEngine.Abstractions
 
         IFile FindBestHostTemplateConfigFile(IFileSystemInfo config);
     }
-
-    public interface IEngineEnvironmentSettings
-    {
-        ISettingsLoader SettingsLoader { get; }
-
-        ITemplateEngineHost Host { get; }
-
-        IEnvironment Environment { get; }
-
-        IPathInfo Paths { get; }
-    }
 }
