+++ src/Microsoft.TemplateEngine.Core/Util/TokenTrie.cs
@@ -87,60 +87,4 @@ namespace Microsoft.TemplateEngine.Core.Util
             return false;
         }
     }
-
-    public class TokenTrieEvaluator : TrieEvaluator<Token>, ITokenTrieEvaluator
-    {
-        private int _currentSequenceNumber;
-
-        public TokenTrieEvaluator(Trie<Token> trie)
-            : base(trie)
-        {
-        }
-
-        public int BytesToKeepInBuffer => _currentSequenceNumber - OldestRequiredSequenceNumber + 1;
-
-        public bool Accept(byte data, ref int bufferPosition, out int token)
-        {
-            ++_currentSequenceNumber;
-            if(Accept(data, ref _currentSequenceNumber, out TerminalLocation<Token> terminal))
-            {
-                token = terminal.Terminal.Index;
-                bufferPosition += _currentSequenceNumber - terminal.Location - terminal.Terminal.End;
-                return true;
-            }
-
-            token = -1;
-            return false;
-        }
-
-        public bool TryFinalizeMatchesInProgress(ref int bufferPosition, out int token)
-        {
-            FinalizeMatchesInProgress(ref _currentSequenceNumber, out TerminalLocation<Token> terminal);
-
-            if(terminal != null)
-            {
-                token = terminal.Terminal.Index;
-                bufferPosition += _currentSequenceNumber - terminal.Location - terminal.Terminal.End;
-                return true;
-            }
-
-            token = -1;
-            return false;
-        }
-    }
-
-    public class Token : TerminalBase
-    {
-        public Token(byte[] token, int index, int start = 0, int end = -1)
-        {
-            Value = token;
-            Index = index;
-            Start = start;
-            End = end != -1 ? end : token.Length;
-        }
-
-        public byte[] Value { get; }
-
-        public int Index { get; }
-    }
 }
