+++ src/Microsoft.TemplateEngine.Edge/ReflectionLoadProbingPath.cs
@@ -10,32 +10,6 @@ using System.Runtime.Loader;
 
 namespace Microsoft.TemplateEngine.Edge
 {
-    public static class TypeEx
-    {
-        public static Type GetType(this string typeName)
-        {
-            int commaIndex = typeName.IndexOf(',');
-            if (commaIndex < 0)
-            {
-                return Type.GetType(typeName);
-            }
-
-            string asmName = typeName.Substring(commaIndex + 1).Trim();
-
-            if (!ReflectionLoadProbingPath.HasLoaded(asmName))
-            {
-                AssemblyName name = new AssemblyName(asmName);
-#if !NET45
-                AssemblyLoadContext.Default.LoadFromAssemblyName(name);
-#else
-                AppDomain.CurrentDomain.Load(name);
-#endif
-            }
-
-            return Type.GetType(typeName);
-        }
-    }
-
     public class ReflectionLoadProbingPath
     {
         private static readonly ConcurrentDictionary<string, Assembly> LoadedAssemblies = new ConcurrentDictionary<string, Assembly>(StringComparer.OrdinalIgnoreCase);
