+++ src/Microsoft.TemplateEngine.Edge/Settings/TemplateInfoReaders/TemplateInfoReaderVersion1_0_0_0.cs
@@ -4,7 +4,7 @@ using Microsoft.TemplateEngine.Abstractions;
 using Microsoft.TemplateEngine.Utils;
 using Newtonsoft.Json.Linq;
 
-namespace Microsoft.TemplateEngine.Edge.Settings
+namespace Microsoft.TemplateEngine.Edge.Settings.TemplateInfoReaders
 {
     public static class TemplateInfoReaderVersion1_0_0_0
     {
