+++ src/Microsoft.TemplateEngine.Edge/Template/TemplateEqualityComparer.cs
@@ -20,18 +20,4 @@ namespace Microsoft.TemplateEngine.Edge.Template
     }
 
     // Compares the templates, irrespective of the match result
-    public class FilteredTemplateEqualityComparer : IEqualityComparer<IFilteredTemplateInfo>
-    {
-        public static IEqualityComparer<IFilteredTemplateInfo> Default { get; } = new FilteredTemplateEqualityComparer();
-
-        public bool Equals(IFilteredTemplateInfo x, IFilteredTemplateInfo y)
-        {
-            return ReferenceEquals(x.Info, y.Info) || (x != null && y != null && x.Info != null && y.Info != null && string.Equals(x.Info.Identity, y.Info.Identity, StringComparison.Ordinal));
-        }
-
-        public int GetHashCode(IFilteredTemplateInfo obj)
-        {
-            return obj?.Info?.Identity?.GetHashCode() ?? 0;
-        }
-    }
 }
