+++ src/Microsoft.TemplateEngine.Orchestrator.RunnableProjects/IReplacementContext.cs
@@ -1,0 +1,9 @@
+namespace Microsoft.TemplateEngine.Orchestrator.RunnableProjects
+{
+    public interface IReplacementContext
+    {
+        string OnlyIfBefore { get; }
+
+        string OnlyIfAfter { get; }
+    }
+}
